refactor(database/offline/bigquery): move functions from helper.go
