+++ internal/database/offline/bigquery/join.go
@@ -2,6 +2,7 @@ package bigquery
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/oom-ai/oomstore/internal/database/offline/sqlutil"
 
@@ -80,3 +81,22 @@ func dropTemporaryTables(ctx context.Context, db *bigquery.Client, tableNames []
 	}
 	return err
 }
+
+func dropTable(ctx context.Context, db *bigquery.Client, tableName string) error {
+	query := fmt.Sprintf(`DROP TABLE IF EXISTS %s;`, tableName)
+	_, err := db.Query(query).Read(ctx)
+	return err
+}
+
+const READ_JOIN_RESULT_QUERY = `
+SELECT
+	{{ qt .EntityRowsTableName }}.{{ .EntityKeyStr }},
+	{{ qt .EntityRowsTableName }}.{{ .UnixMilliStr }},
+	{{ fieldJoin .Fields }}
+FROM {{ $.DatasetID }}.{{ qt .EntityRowsTableName }}
+{{ range $pair := .JoinTables }}
+	{{- $t1 := qt $pair.LeftTable -}}
+	{{- $t2 := qt $pair.RightTable -}}
+lEFT JOIN {{ $.DatasetID }}.{{ $t2 }}
+ON {{ $t1 }}.{{ $.UnixMilliStr }} = {{ $t2 }}.{{ $.UnixMilliStr }} AND {{ $t1 }}.{{ $.EntityKeyStr }} = {{ $t2 }}.{{ $.EntityKeyStr }}
+{{end}}`
